Adding sharing - needed to modify quite a few files, like eoDistance.h make_algo_scalar.h and all related files, and the like

+++ eo/tutorial/Lesson4/binary_value.h
@@ -7,9 +7,9 @@
     @param _chrom A binary chromosome 
 */
 
-template <class Chrom> float binary_value(const Chrom& _chrom)
+template <class Chrom> double binary_value(const Chrom& _chrom)
 {
-  float sum = 0;
+  double sum = 0;
   for (unsigned i = 0; i < _chrom.size(); i++)
     if (_chrom[i])
 	sum += _chrom[i];
